Move registration of ImportOrganizerRequestHandler to OrganizerControllerProvider
