+++ app/Import/ImportControllerProvider.php
@@ -4,11 +4,8 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Silex\Import;
 
-use CultuurNet\UDB3\Http\Import\ImportLabelVisibilityRequestBodyParser;
 use CultuurNet\UDB3\Http\Import\ImportRestController;
 use CultuurNet\UDB3\Http\Organizer\ImportOrganizerRequestHandler;
-use CultuurNet\UDB3\Http\Request\Body\CombinedRequestBodyParser;
-use CultuurNet\UDB3\Silex\Labels\LabelServiceProvider;
 use Silex\Application;
 use Silex\ControllerCollection;
 use Silex\ControllerProviderInterface;
@@ -41,22 +38,6 @@ class ImportControllerProvider implements ControllerProviderInterface
                     $app['place_iri_generator']
                 );
             }
-        );
-
-        $app[ImportOrganizerRequestHandler::class] = $app->share(
-            fn (Application $app) => new ImportOrganizerRequestHandler(
-                $app['organizer_repository'],
-                $app['imports_command_bus'],
-                $app['labels.labels_locked_for_import_repository'],
-                $app['uuid_generator'],
-                $app['organizer_iri_generator'],
-                new CombinedRequestBodyParser(
-                    new ImportLabelVisibilityRequestBodyParser(
-                        $app[LabelServiceProvider::JSON_READ_REPOSITORY],
-                        $app[LabelServiceProvider::RELATIONS_READ_REPOSITORY]
-                    )
-                )
-            )
         );
 
         /* @var ControllerCollection $controllers */
